+++ aiobungie/__init__.py
@@ -96,9 +96,11 @@ from aiobungie import crate
 from aiobungie import interfaces
 from aiobungie import traits
 from aiobungie import typedefs
+from aiobungie import url
 from aiobungie.client import Client
 from aiobungie.error import *
 from aiobungie.internal.enums import *
+from aiobungie.rest import OAuth2Response
 from aiobungie.rest import RESTClient
 from aiobungie.undefined import Undefined
 from aiobungie.undefined import UndefinedOr